requirements.txt
streamlit
requests
pandas
numpy
pillow
openai
pyyaml
python-dotenv
librosa
emoji